streamlined special cart detection
