+++ src/cart/cart.cpp
@@ -2,6 +2,7 @@
 #include "frontend/logger.hpp"
 #include <algorithm>
 #include <fstream>
+#include <bitset>
 
 static std::optional<std::vector<byte_t>> read_all_bytes(const fs::path& p) {
   std::ifstream f(p, std::ios::binary | std::ios::ate);
@@ -190,7 +191,7 @@ bool validate(cart& c) {
       fail_count += 1;
     }
     else {
-      if (offset != 0) c.is_mmm01 = true;
+      if (offset != 0) c.special_mbc = MMM01;
       // If declared size is known, ensure file is at least that big
       if (c.declared_rom_bytes != 0 && c.rom.size() < c.declared_rom_bytes) {
         Logger::push(LogLevel::Warning, "ROM", "ROM too small",
@@ -207,6 +208,94 @@ bool validate(cart& c) {
   return true;
 }
 
+/* Weird carts detection */
+// Wisdom Tree detection because it's autistic :(
+bool maybe_wisdom_tree(const std::span<const byte_t> rom) {
+  if (rom.size() <= 0x8000) return false;
+
+  // scan the first chunk to catch the init code
+  const std::size_t limit = std::min<std::size_t>(rom.size(), 0x40000); // 256 KiB
+
+  std::uint32_t ea_total = 0;
+  std::uint32_t ea_cart  = 0;
+  std::bitset<256> low_bytes{};
+
+  for (std::size_t i = 0; i + 2 < limit; ++i) {
+    if (rom[i] != 0xEA) continue; // LD (a16),A
+
+    ++ea_total;
+    const auto lo = static_cast<std::uint8_t>(rom[i + 1]);
+    const auto hi = static_cast<std::uint8_t>(rom[i + 2]);
+    // address = hi<<8 | lo
+    if (hi < 0x80) { // 0000-7FFF (cartridge / mapper control area)
+      ++ea_cart;
+      low_bytes.set(lo);
+    }
+  }
+  const std::size_t distinct_lo = low_bytes.count();
+
+  // Heuristic thresholds:
+  // - need some evidence of cart-area stores
+  // - need multiple distinct low bytes (since WT bank is low byte of address)
+  // - and a decent fraction of EA stores going to cart area
+  if (ea_cart < 8) return false;
+  if (distinct_lo < 6) return false;
+
+  if (ea_total > 0) {
+    const double frac = static_cast<double>(ea_cart) / static_cast<double>(ea_total);
+    if (frac < 0.35) return false;
+  }
+
+  return true;
+}
+
+// Noooo, not you M161 too :(
+bool maybe_m161(const std::span<const byte_t> rom) {
+  // M161 maps 32 KiB banks into 0000-7FFF, bank number is 3 bits (00-07)
+  // So if the ROM is bigger than 32 KiB but doesn't have a whole number of 32 KiB banks, it's likely not M161
+  // HOWEVER this is still not very foolproof. Need more research...
+  if (rom.size() <= 0x8000) return false;
+  if (rom.size() % 0x8000 != 0) return false;          // whole number of 32 KiB banks
+  if (rom.size() > 0x8000 * 8) return false;           // max 8 banks
+
+  return true;
+}
+
+SpecialMbc detect_special_mbc(const cart& c) {
+  switch (c.header.cartridge_type) {
+  case 0x00: { // ROM ONLY, but some WT/M161 carts lie about this, we investigate further
+    if (c.rom_size() <= 0x8000) return None;  // If strictly <= 32KiB, it's probably safe
+    if (c.header.title() == "WISDOM TREE" || maybe_wisdom_tree(c.rom_span())) {
+      Logger::push(
+        LogLevel::Warning, "ROM", "Mapper override",
+      std::format("{} header type {:02X} looks inconsistent with ROM size {} and appears to be WT; "
+                  "forcing Wisdom Tree mapper.",
+                  c.header.title(), c.header.cartridge_type, c.rom_span().size()));
+      return WisdomTree;
+    }
+    if (maybe_m161(c.rom_span())) {
+      Logger::push(
+        LogLevel::Warning, "ROM", "Mapper override",
+      std::format("{} header type {:02X} looks inconsistent with ROM size {} and appears to be M161; "
+                  "forcing M161 mapper.",
+                  c.header.title(), c.header.cartridge_type, c.rom_span().size()));
+      return M161;
+    }
+  }
+  case 0x0F: case 0x10: case 0x11: case 0x12: case 0x13:  // Special MBC3 that has 64 KiB RAM
+    if (c.declared_ram_bytes > 32 * 1024)
+      return MBC30;
+  case 0x1B:
+    if (c.header.destination_code == 0xE1 || c.header.title() == "EMSMENU" || c.header.title() == "GB16M")
+      return EMS;
+  case 0xC0:
+    if (c.header.destination_code == 0xD1)
+      return WisdomTree;
+  default:
+    return None;
+  }
+}
+
 std::optional<cart> load_cart_fs(const fs::path& rom_path) {
   cart c{};
   c.file_path = rom_path;
@@ -215,7 +304,10 @@ std::optional<cart> load_cart_fs(const fs::path& rom_path) {
   else
     return std::nullopt;
 
-  if (validate(c)) return c;
+  if (validate(c)) {
+    c.special_mbc = detect_special_mbc(c);
+    return c;
+  }
 
   return std::nullopt;
 }
@@ -225,5 +317,6 @@ cart load_cart_raw(std::vector<byte_t> rom_bytes) {
   c.file_path.clear();
   c.rom = std::move(rom_bytes);
   validate(c);
+  c.special_mbc = detect_special_mbc(c);
   return c;
 }
