+++ src/cart/mbc/mbc3.cpp
@@ -248,9 +248,7 @@ std::unique_ptr<Mbc> make_mbc3(const cart &c) {
   const bool has_rtc =
       c.header.cartridge_type == 0x0F || // MBC3+TIMER+BATTERY
       c.header.cartridge_type == 0x10;  // MBC3+TIMER+RAM+BATTERY
-  const bool is_mbc30 =
-    c.rom_size() > 2 * 1024 * 1024 || c.declared_ram_bytes > 32 * 1024;
   return std::make_unique<Mbc3>(c.rom_span(), c.declared_ram_bytes,
                                 type_has_battery(c.header.cartridge_type),
-                                has_rtc, is_mbc30);
+                                has_rtc, c.special_mbc == MBC30);
 }
